sesion_05/requirements.txt
# Django
Django==3.1.3

# REST Framework
djangorestframework==3.12.1
django-filter==2.4.0
Markdown==3.3.3

# JWT
djangorestframework-simplejwt==4.5.0

# CORS
django_cors_headers==3.5.0
